refactored enough that the demo scripts seem relatively sane and readable
