+++ rag.py
@@ -8,38 +8,14 @@ based on BAML <https://docs.boundaryml.com/examples/prompt-engineering/retrieval
 see copyright/license https://github.com/DerwenAI/strwythura/README.md
 """
 
-import json
 import logging
 import os
 import traceback
-import typing
 import warnings
 
 from icecream import ic
 
 from strwythura import Strwythura, GraphRAG
-from strwythura import baml_client
- 
-
-def qa_loop (
-    rag: GraphRAG,
-    ) -> None:
-    """
-Loop to answer questions.
-    """
-    while True:
-        question: str = input("\n\nWhat is your question? ").strip()
-
-        if question.lower() in [ "exit", "quit" ]:
-            print("\nÀ bientôt!")
-            break
-
-        chunks: typing.List[ str ] = rag.get_chunks(question, debug = False) # True
-        context: str = "\n".join( chunks )
-        response: baml_client.types.Response = baml_client.b.RAG(question, context)
-
-        ic(response)
-        print("-" * 10)
     
 
 if __name__ == "__main__":
@@ -62,9 +38,9 @@ if __name__ == "__main__":
 
         rag: GraphRAG = GraphRAG(strw)
 
-        # loop to answer questions
         try:
-            qa_loop(rag)
+            # loop to answer questions
+            rag.qa_loop(debug = False) # True
 
         except EOFError:
             print("")
